requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
sqlalchemy==2.0.0
aiosqlite==0.19.0
alembic==1.12.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
pydantic-settings==2.1.0 